some exception handling, adjustments

+++ src/main/java/mwisnie/project/airqualitymonitorbackend/controller/UserController.java
@@ -2,13 +2,16 @@ package mwisnie.project.airqualitymonitorbackend.controller;
 
 import lombok.RequiredArgsConstructor;
 import mwisnie.project.airqualitymonitorbackend.entity.User;
-import mwisnie.project.airqualitymonitorbackend.listener.OnRegistrationEvent;
+import mwisnie.project.airqualitymonitorbackend.event.OnRegistrationEvent;
+import mwisnie.project.airqualitymonitorbackend.exception.types.RegistrationException;
+import mwisnie.project.airqualitymonitorbackend.exception.types.UnauthorizedException;
 import mwisnie.project.airqualitymonitorbackend.service.user.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.context.request.WebRequest;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -23,37 +26,55 @@ public class UserController {
     private final ApplicationEventPublisher applicationEventPublisher;
 
     @GetMapping
-    public List<User> getAllUsers() {
-        return userService.getAllUsers();
+    public List<User> getAllUsers(Authentication auth) {
+        if (isAdmin(auth)) {
+            return userService.getAllUsers();
+        }
+        return new ArrayList<>();
     }
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable("id") String id) {
+    public User getUserById(@PathVariable("id") String id, Authentication auth) {
+        if (!isAuthorized(id, auth)) {
+            throw new UnauthorizedException(String.format("Not authorized to see data of user %s.", id));
+        }
         return userService.getUserById(id);
     }
 
     @PostMapping
-    public User createUser(@RequestBody User user, WebRequest request) {
-        //if username exists, throw throw error
-
+    public User createUser(@RequestBody User user) {
         User createdUser = userService.createUser(user);
         if (createdUser == null) {
-            //todo: throw specific error
-            return null;
+            throw new RegistrationException("Error - unable to create the user.");
         }
 
         applicationEventPublisher.publishEvent(new OnRegistrationEvent(createdUser));
         return createdUser;
     }
 
-    @PutMapping
-    public User updateUser(@RequestBody User user) {
+    @PutMapping("/{id}")
+    public User updateUser(@PathVariable("id") String id, @RequestBody User user, Authentication auth) {
+        if (!isAuthorized(id, auth)) {
+            throw new UnauthorizedException(String.format("Not authorized to edit user %s.", id));
+        }
         return userService.updateUser(user);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable String id) {
+    public void deleteUser(@PathVariable String id, Authentication auth) {
+        if (!isAuthorized(id, auth)) {
+            throw new UnauthorizedException(String.format("Not authorized to delete user %s.", id));
+        }
         userService.deleteUserById(id);
+    }
+
+    private boolean isAuthorized(String id, Authentication auth) {
+        return id.equals(userService.getUserByUsername(auth.getPrincipal().toString()).getId());
+    }
+
+    private boolean isAdmin(Authentication auth) {
+        //todo: implementation
+        return false;
     }
 
 }
